refactor: 💡 Fixed all CR Comments
destructuring params directly in ngNewArgsFromDto, renamed to
GENERATE_COMMAND, renamed to CREATE_WORKSPACE_COMMAND, created
commands.ts, removed base class, added a comment to explain execAsync in
createWorkspace, provided GeneratorsService in workspace.controller.spec.ts
