+++ apps/cli-daemon/src/app/generators/generators.module.ts
@@ -6,5 +6,6 @@ import { GeneratorsService } from './generators.service';
 @Module({
   controllers: [ComponentController],
   providers: [GeneratorsService],
+  exports: [GeneratorsService],
 })
 export class GeneratorsModule {}
