+++ apps/cli-daemon/src/app/generators/generators.service.ts
@@ -1,15 +1,26 @@
+import { spawnSync, SpawnSyncOptionsWithBufferEncoding } from 'child_process';
+import { resolve as pathResolve } from 'path';
+
 import { Injectable } from '@nestjs/common';
 
-import { NgCommandExecutorBase } from '../abstract/ng-command-executor.base';
+import { GENERATE_COMMAND } from '../ng-commands';
 import { SessionService } from '../session/session.service';
 
-export const NG = 'npx ng g';
+import { ExecResult } from './dto';
 
 @Injectable()
-export class GeneratorsService extends NgCommandExecutorBase {
-  readonly ng = NG;
+export class GeneratorsService {
+  constructor(private readonly sessionService: SessionService) {}
 
-  constructor(sessionService: SessionService) {
-    super(sessionService);
+  execSync(args: string[] = [], ng = GENERATE_COMMAND): ExecResult {
+    const options: SpawnSyncOptionsWithBufferEncoding = {
+      cwd: pathResolve(this.sessionService.cwd),
+      env: process.env,
+      shell: true,
+    };
+    const command = [ng].concat(args || []).join(' ');
+    const res = spawnSync(command, options);
+
+    return new ExecResult(res);
   }
 }
