+++ apps/cli-daemon/src/app/ng-commands.ts
@@ -1,0 +1,2 @@
+export const GENERATE_COMMAND = 'npx ng g';
+export const CREATE_WORKSPACE_COMMAND = 'npx ng new';
