+++ apps/cli-daemon/src/app/workspace/workspace.module.ts
@@ -1,9 +1,12 @@
 import { Module } from '@nestjs/common';
 
+import { GeneratorsModule } from '../generators/generators.module';
+
 import { WorkspaceController } from './workspace.controller';
 import { WorkspaceService } from './workspace.service';
 
 @Module({
+  imports: [GeneratorsModule],
   providers: [WorkspaceService],
   controllers: [WorkspaceController],
 })
