+++ apps/cli-daemon/src/app/workspace/workspace.service.ts
@@ -14,7 +14,6 @@ import {
   NotFoundException,
 } from '@nestjs/common';
 
-import { NgCommandExecutorBase } from '../abstract/ng-command-executor.base';
 import { SessionService } from '../session/session.service';
 
 import {
@@ -26,14 +25,10 @@ import {
 export const ng = 'npx ng new';
 
 @Injectable()
-export class WorkspaceService extends NgCommandExecutorBase {
-  readonly ng = ng;
-
+export class WorkspaceService {
   private readonly logger = new Logger(WorkspaceService.name);
-
-  constructor(sessionService: SessionService) {
-    super(sessionService);
-  }
+  readonly ng = ng;
+  constructor(private sessionService: SessionService) {}
 
   async readWorkspace(path: string): Promise<WorkspaceDefinition> {
     try {
